Delete BooBuddies/env directory
environments are abs pathed to your machine. you must make a new environment using the requirements.txt. in vs code click at the top and add a > symbol. click on python create environment, then select Venv then if the venv exists use delete and recreate or if it doesnt exist use create, make sure you click on your machines distribution of the python interpreter the other will have a path to microsoft so dont use that one. next it will ask for dependencies simply check the requirements.txt check box and it will create the environment for you.